Add the authentication for the management of the store

+++ resources/views/products/show.blade.php
@@ -8,9 +8,11 @@
         <p>Created {{ $product->created_at->diffForHumans() }}</p>
 
         <a href="{{ route('products.index') }}" class="btn btn-primary button-green pull-right">Retourner au magasin</a>
-        <a href="{{ route('products.edit', $product) }}" class="btn btn-primary pull-right">Editer le produuit</a>
-        {!! Form::open(['method' => 'DELETE', 'route' => ['products.destroy', $product]]) !!}
-        {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right'] ) !!}
-        {!! Form::close() !!}
+        @if(Auth::check() && Auth::user()->status===1)
+            <a href="{{ route('products.edit', $product) }}" class="btn btn-primary pull-right">Editer le produuit</a>
+            {!! Form::open(['method' => 'DELETE', 'route' => ['products.destroy', $product]]) !!}
+            {!! Form::submit('Supprimer', ['class' => 'btn btn-danger pull-right'] ) !!}
+            {!! Form::close() !!}
+        @endif
     </div>
 @endsection
